Count Introduction and Summary as chapters
I don't remember my justification when I introduced these lines, but
nowadays it seems that most people expect the total number of chapters
to be equal to the number of actual chapters. So if Introduction and
Summary are numbered, then they're counted as chapters. And they're
numbered in the main template, so they will be numbered here too (at
least for now).